% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPackages.R
\name{checkPackages}
\alias{checkPackages}
\title{checkPackages}
\usage{
checkPackages(
  ...,
  add = c("highlight", "formatR", "shiny", "shinydashboard", "shinydashboardPlus", "DT")
)
}
\arguments{
\item{...}{character: name(s) of package}

\item{add}{character: names of default packages to check (default: \code{c("highlight", "formatR", "shiny", "shinydashboard", "shinydashboardPlus", "DT")})}
}
\value{
\code{TRUE} if successful otherweise an error will be thrown
}
\description{
Checks if a package is installed without loading it. Returns a logical vector with \code{TRUE} or \code{FALSE} for each package checked.
}
\examples{
checkPackages("graphics", add=NULL)          # checks if 'graphics' is installed
if (interactive()) checkPackages("graphics") # checks if 'graphics', 'shiny', ... are installed
}
