% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionParameters.R
\name{distributionParams}
\alias{distributionParams}
\title{distributionParams}
\usage{
distributionParams(mean, sd, n = 30, N = 60)
}
\arguments{
\item{mean}{numeric: mean}

\item{sd}{numeric: standard deviation (only used for the normal distribution)}

\item{n}{integer: sample size (default: \code{30})}

\item{N}{integer: population size (default: \code{60})}
}
\value{
a list of parameters for each distribution
}
\description{
Computes approximate distribution parameters for the binomial, hypergeometric,
Poisson, Exponential and normal distribution for a given mean (and standard deviation).
With the sample and the population size the computation can be influenced.
}
\examples{
# Compute approx. paramaters for a binomial distribution
distributionParams(mean=30*0.5, sd=sqrt(30*0.5*0.5))
}
