% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gettext.R
\name{gettext}
\alias{gettext}
\title{gettext}
\usage{
gettext(msg, utype = "vector")
}
\arguments{
\item{msg}{character: message(s) to translate}

\item{utype}{character: how to return the translated message as vector or named list}
}
\value{
translated messages
}
\description{
Returns a translation from loaded PO-file. If the message is not found in the PO-file then original text will be returned.
}
\examples{
msgs <- c("two.sided", "less", "greater")
gettext(msgs)
# for use in Shiny "choices"
gettext(msgs, "name")
gettext(msgs, "numeric")
}
