% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.ASCII.R
\name{is.ASCII}
\alias{is.ASCII}
\title{is.ASCII}
\usage{
is.ASCII(txt)
}
\arguments{
\item{txt}{character: text to check}
}
\value{
logical
}
\description{
Checks if \code{txt} contains only ASCII characters.
}
\examples{
is.ASCII("Congratulations")
is.ASCII("Herzlichen Glückwunsch")
}
