% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_asso.R
\name{men_asso}
\alias{men_asso}
\title{Association}
\usage{
men_asso(...)
}
\arguments{
\item{...}{one or more data sets}
}
\value{
nothing
}
\description{
Shiny app for the association coefficients between two categorical variables.
If no data are given then the default data from the book will be used.
Otherwise the data will be stored as RDS file in a temporary directory.
}
\examples{
if (interactive()) men_asso()
if (interactive()) men_asso(HairEyeColor, Titanic)
}
