% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_bin.R
\name{men_bin}
\alias{men_bin}
\title{men_bin}
\usage{
men_bin(size = 10, prob = 0.5)
}
\arguments{
\item{size}{integer: number of trials (zero or more)}

\item{prob}{numeric: probability of success on each trial}
}
\value{
nothing
}
\description{
Visualization of the probability mass and the cumulative distribution function of
a binomial distribution.
}
\examples{
if (interactive()) men_bin()
if (interactive()) men_bin(20, 0.25)
}
