% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_cilen.R
\name{men_cilen}
\alias{men_cilen}
\title{men_cilen}
\usage{
men_cilen()
}
\value{
nothing
}
\description{
Shiny app for a length of a confidence interval for the mean.
}
\examples{
if (interactive()) men_cilen()
}
