% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_cipi.R
\name{men_cipi}
\alias{men_cipi}
\title{men_cipi}
\usage{
men_cipi(...)
}
\arguments{
\item{...}{one or more data sets}
}
\value{
nothing
}
\description{
Shiny app for a confidence interval for the proportion.
The data used is considered as a population from which random samples can be drawn.
If no data are given then the default data from the book will be used.
Otherwise the data will be stored as RDS file in a temporary directory.
}
\examples{
if (interactive()) men_cipi()
if (interactive()) men_cipi(Titanic)
}
