% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_corr.R
\name{men_corr}
\alias{men_corr}
\title{Correlation}
\usage{
men_corr(...)
}
\arguments{
\item{...}{one or more data sets}
}
\value{
nothing
}
\description{
Shiny app for the correlation coefficients between two numeric variables.
If no data are given then the default data from the book will be used.
Otherwise the data will be stored as RDS file in a temporary directory.
}
\examples{
if (interactive()) men_corr()
if (interactive()) men_corr(iris)
}
