% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_exp.R
\name{men_exp}
\alias{men_exp}
\title{men_exp}
\usage{
men_exp(rate = 1)
}
\arguments{
\item{rate}{numeric: rate}
}
\value{
nothing
}
\description{
Visualization of the density and the cumulative distribution function of
a exponential distribution.
}
\examples{
if (interactive()) men_exp()
if (interactive()) men_exp(3)
}
