% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_hall.R
\name{men_hall}
\alias{men_hall}
\title{men_hall}
\usage{
men_hall(pointdoor = 1, afteropen = 1)
}
\arguments{
\item{pointdoor}{integer: to which door to point (default: \code{1})}

\item{afteropen}{integer: play strategy 1=keep door, 2=change door (default: \code{1})}
}
\value{
nothing
}
\description{
Shiny app for the \href{https://en.wikipedia.org/wiki/Monty_Hall_problem}{Monty Hall problem}:
}
\details{
Suppose you're on a game show, and you're given the choice of three doors: Behind one door is a car; behind the others, goats. You pick a door, say No. 1, and the host, who knows what's behind the doors, opens another door, say No. 3, which has a goat. He then says to you, "Do you want to pick door No. 2?" Is it to your advantage to switch your choice?
}
\examples{
if (interactive()) men_hall()
if (interactive()) men_hall(4, 2)
}
