% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_hist.R
\name{men_hist}
\alias{men_hist}
\title{men_hist}
\usage{
men_hist(...)
}
\arguments{
\item{...}{one or more data sets}
}
\value{
nothing
}
\description{
Shiny app for visualizing a univariate numeric variable as histögram.
If no data are given then the default data from the book will be used.
Otherwise the data will be stored as RDS file in a temporary directory.  #'
}
\examples{
if (interactive()) men_hist()
if (interactive()) men_hist(iris)
}
