% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_poi.R
\name{men_poi}
\alias{men_poi}
\title{men_poi
Visualization of the probability mass and the cumulative distribution function of
a Poisson distribution.}
\usage{
men_poi(lambda = 5)
}
\arguments{
\item{lambda}{numeric: (non-negative) mean}
}
\value{
nothing
}
\description{
men_poi
Visualization of the probability mass and the cumulative distribution function of
a Poisson distribution.
}
\examples{
if (interactive()) men_poi()
if (interactive()) men_poi(3)
}
