% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_rank.R
\name{men_rank}
\alias{men_rank}
\title{men_rank}
\usage{
men_rank(...)
}
\arguments{
\item{...}{one or more data sets}
}
\value{
nothing
}
\description{
Shiny app for the rank correlation coefficients between two ordered variables.
If no data are given then the default data from the book will be used.
Otherwise the data will be stored as RDS file in a temporary directory.
}
\examples{
if (interactive()) men_rank()
if (interactive()) {
  data("plantTraits", package="cluster")
  men_rank(plantTraits)
}
}
