% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_regr.R
\name{men_regr}
\alias{men_regr}
\title{men_regr}
\usage{
men_regr(...)
}
\arguments{
\item{...}{one or more data sets}
}
\value{
nothing
}
\description{
Shiny app for a simple linear regression.
If no data are given then the default data from the book will be used.
Otherwise the data will be stored as RDS file in a temporary directory.
}
\examples{
if (interactive()) men_regr()
if (interactive()) men_regr(stackloss)
}
