% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_tab.R
\name{men_tab}
\alias{men_tab}
\title{Frequency tables}
\usage{
men_tab(...)
}
\arguments{
\item{...}{one or more data sets}
}
\value{
nothing
}
\description{
Shiny app for frequency tables for two categorical variables.
If no data are given then the default data from the book will be used.
Otherwise the data will be stored as RDS file in a temporary directory.
}
\examples{
if (interactive()) men_tab()
if (interactive()) men_tab(HairEyeColor, Titanic)
}
