% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/men_vis.R
\name{men_vis}
\alias{men_vis}
\title{men_vis}
\usage{
men_vis(...)
}
\arguments{
\item{...}{one or more data sets}
}
\value{
nothing
}
\description{
Shiny app for visualizing the univariate numeric  variable, e.g. boxplot, stripchart, histogram,
and cumulative distribution function.
If no data are given then the default data from the book will be used.
Otherwise the data will be stored as RDS file in a temporary directory.
}
\examples{
if (interactive()) men_vis()
if (interactive()) men_vis(iris)
}
