% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.getDataNames.R
\name{mmstat.getDataNames}
\alias{mmstat.getDataNames}
\title{mmstat.getDataNames}
\usage{
mmstat.getDataNames(...)
}
\arguments{
\item{...}{character: names of the data sets.}
}
\value{
the names of the data sets
}
\description{
Returns the names of data sets and stores them in the internal environment.
The name of the data set is base name without extension.
}
\examples{
files <- mmstat.rds("HAIR.EYE.COLOR", "TITANIC")
mmstat.getDataNames(files)
}
