% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.getDatasets.R
\name{mmstat.getDatasets}
\alias{mmstat.getDatasets}
\title{mmstat.getDatasets}
\usage{
mmstat.getDatasets(...)
}
\arguments{
\item{...}{character: file name(s) of RDS data file(s)}
}
\value{
the names of the data set(s)
}
\description{
Reads data set(s) into the mmstat object.
}
\examples{
# not used, deprecated??
}
