% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.getLog.R
\name{mmstat.getLog}
\alias{mmstat.getLog}
\title{mmstat.getLog}
\usage{
mmstat.getLog(session)
}
\arguments{
\item{session}{session object}
}
\value{
HTML code
}
\description{
Returns the internal log message as HTML. In a Shiny app the log message are updated every 100 milliseconds
}
\examples{
# will work only in Shiny app:
# mmstat.getLog(session)
}
