% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.getValues.R
\name{mmstat.getValues}
\alias{mmstat.getValues}
\alias{mmstat.getValue}
\title{mmstat.getValues}
\usage{
mmstat.getValues(local, ...)

mmstat.getValue(val, def)
}
\arguments{
\item{local}{list: default values for the named parameter}

\item{...}{list of named parameters}

\item{val}{value for a parameter}

\item{def}{default value for a parameter}
}
\value{
a list of requested parameters
}
\description{
\code{mmstat.getValues} returns a list with named elements. If the parameter is \code{NULL} then a default value stored
\code{local} will be used.
\code{mmstat.getValue} returns a value. If the parameter is \code{NULL} or \code{NA} then \code{def} will be returned
}
\examples{
def <- list(a=3)
mmstat.getValues(def, b=3, a=NULL)
mmstat.getValue(NA, 5)
mmstat.getValue(NULL, 5)
mmstat.getValue(3, 5)
}
