% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.getVariableNames.R
\name{mmstat.getVariableNames}
\alias{mmstat.getVariableNames}
\title{mmstat.getVariableNames}
\usage{
mmstat.getVariableNames(name)
}
\arguments{
\item{name}{character or numeric: name or index of data set}
}
\value{
vector of names
}
\description{
Returns all variable names of data set stored in the internal environment.
}
\examples{
# Delete all stored data sets
mmstat.set(datasets=NULL)
# Load CAR data set into mmstat
mmstat.getDataNames(mmstat.rds("CARS"))
# Extract names of all variables
mmstat.getVariableNames(1)
}
