% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.getVarNames.R
\name{mmstat.getVarNames}
\alias{mmstat.getVarNames}
\title{mmstat.getVarNames}
\usage{
mmstat.getVarNames(dataname, vartype, which = NULL)
}
\arguments{
\item{dataname}{character: name of data set}

\item{vartype}{character: type of variable, either \code{numeric},  \code{ordered},  \code{factor}, or \code{binary}}

\item{which}{integer: index number}
}
\value{
a vector or element of variable names which have the type \code{vartype}
}
\description{
Returns the variable names of a specific type from a mmstat data set.
}
\examples{
# Load CAR data set into mmstat
mmstat.getDataNames(mmstat.rds("CARS"))
# Extract names of numeric variables
mmstat.getVarNames(1, "numeric")
}
