% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.lang.R
\name{mmstat.lang}
\alias{mmstat.lang}
\title{mmstat.lang}
\usage{
mmstat.lang(pof = NULL)
}
\arguments{
\item{pof}{character: file name}
}
\value{
nothing
}
\description{
Loads a \href{https://www.gnu.org/software/gettext/manual/html_node/PO-Files.html}{PO file} for a translation into the internal environment.
}
\examples{
mmstat.lang()
}
