% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.log.R
\name{mmstat.log}
\alias{mmstat.log}
\title{mmstat.log}
\usage{
mmstat.log(txt)
}
\arguments{
\item{txt}{character: message to write}
}
\value{
nothing
}
\description{
Writes a message into the internal log.
}
\examples{
mmstat.log("Test")
}
