% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.math.R
\name{mmstat.math}
\alias{mmstat.math}
\title{mmstat.math}
\usage{
mmstat.math(txt)
}
\arguments{
\item{txt}{character: input text}
}
\value{
expression
}
\description{
Returns a math expression based on HTML special characters notation.
}
\examples{
mmstat.math(" &bar(X);~&N(mu[0], sigma^2/n); ")
mmstat.math("&H[0];: &mu==mu[0]; vs. &H[1];: &mu!=mu[0]; ")
}
