% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.merge.R
\name{mmstat.merge}
\alias{mmstat.merge}
\title{mmstat.merge}
\usage{
mmstat.merge(range1, range2)
}
\arguments{
\item{range1}{range: first range}

\item{range2}{range: second range}
}
\value{
new range
}
\description{
Computes a new range from by a union of the two ranges.
}
\examples{
mmstat.merge(c(0,1), c(0.5, 2)) # returns c(0, 2)
}
