% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.pos.R
\name{mmstat.pos}
\alias{mmstat.pos}
\title{mmstat.pos}
\usage{
mmstat.pos(minmax, pos)
}
\arguments{
\item{minmax}{numeric(2): range to interpolate between}

\item{pos}{numeric: proportion(s) to interpolate, usually between zero and one}
}
\value{
interpolated values
}
\description{
Returns a linear interpolation based on \code{minmax}.
}
\examples{
mmstat.pos(c(0,360), 0.5)
}
