% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.range.R
\name{mmstat.range}
\alias{mmstat.range}
\title{mmstat.range}
\usage{
mmstat.range(...)
}
\arguments{
\item{...}{R objects}
}
\value{
range
}
\description{
Computes a range from several R objects by union.
}
\examples{
mmstat.range(-5:5, 0:10) # returns c(-5, 10)
}
