% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.rds.R
\name{mmstat.rds}
\alias{mmstat.rds}
\title{mmstat.rds}
\usage{
mmstat.rds(...)
}
\arguments{
\item{...}{names of data sets}
}
\value{
full file names
}
\description{
Returns the full file names of all or specific data set that come with the package.
}
\examples{
mmstat.rds()                            # return all RDS file that come with the package
mmstat.rds("HAIR.EYE.COLOR", "TITANIC") # location of specific data sets
}
