% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.round.down.R
\name{mmstat.round.down}
\alias{mmstat.round.down}
\title{mmstat.round.down}
\usage{
mmstat.round.down(x, digits = 0)
}
\arguments{
\item{x}{numeric: values for rounding}

\item{digits}{numeric: digits for rounding (default: \code{0})}
}
\value{
down rounded values
}
\description{
Rounds down.
}
\examples{
x <- runif(5)
cbind(x, mmstat.round.down(x, 1))
}
