% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.round.up.R
\name{mmstat.round.up}
\alias{mmstat.round.up}
\title{mmstat.round.up}
\usage{
mmstat.round.up(x, digits = 0)
}
\arguments{
\item{x}{numeric: values for rounding}

\item{digits}{numeric: digits for rounding (default: \code{0})}
}
\value{
uprounded values
}
\description{
Rounds up.
}
\examples{
x <- runif(5)
cbind(x, mmstat.round.up(x, 1))
}
