% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.set.R
\name{mmstat.set}
\alias{mmstat.set}
\alias{mmstat.get}
\title{mmstat.set}
\usage{
mmstat.set(...)

mmstat.get(...)
}
\arguments{
\item{...}{named parameters with values or names}
}
\value{
nothing
}
\description{
\code{mmstat.set} sets one (or more) parameter to the internal environment. \code{mmstat.get} return one or more parameters from the internal environment.
}
\examples{
mmstat.set(debug=0)
mmstat.get("debug")
mmstat.get("debug", "shiny") # returns a list
}
