% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.sliderInput.R
\name{mmstat.sliderInput}
\alias{mmstat.sliderInput}
\title{mmstat.sliderInput}
\usage{
mmstat.sliderInput(...)
}
\arguments{
\item{...}{parameters for \link[shiny:sliderInput]{shiny::sliderInput}}
}
\value{
the HTML output
}
\description{
A modified \code{sliderInput} for \code{mmstat} which supports user defined tick marks.
}
\examples{
ticks <- c(80, 85, 90, 95, 98, 99, 99.5, 99.9)
mmstat.sliderInput("id", "label", min=1, max=length(ticks), value=3, step=1, ticks=ticks)
}
