% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.ui.call.R
\name{mmstat.ui.call}
\alias{mmstat.ui.call}
\title{mmstat.ui.call}
\usage{
mmstat.ui.call(inputId, ...)
}
\arguments{
\item{inputId}{character: the input slot called}

\item{...}{further parameters given to the call}
}
\value{
whatever the call to the underlying Shiny UI element returns
}
\description{
Calls the underlying Shiny UI element (\code{selectInput}, ...).
}
\examples{
mmstat.ui.elem(inputId="alpha", type="significance")
mmstat.ui.call("alpha")
}
