% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.ui.update.R
\name{mmstat.ui.update}
\alias{mmstat.ui.update}
\title{mmstat.ui.update}
\usage{
mmstat.ui.update(inputId, ...)
}
\arguments{
\item{inputId}{character: the input slot called}

\item{...}{further parameters given to the call}
}
\value{
whatever the update to the underlying Shiny UI element returns
}
\description{
Call for a update of an underlying Shiny UI element (\code{selectInput}, ...).
}
\examples{
mmstat.ui.elem(inputId="alpha", type="significance")
mmstat.ui.call("alpha")
}
