% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmstat.warn.R
\name{mmstat.warn}
\alias{mmstat.warn}
\title{mmstat.warn}
\usage{
mmstat.warn(cond, txt)
}
\arguments{
\item{cond}{logical: condition to test}

\item{txt}{character: text to write if \code{cond} is true}
}
\value{
nothing
}
\description{
Writes a warning text into the log object in the internal mmstat object.
}
\examples{
mmstat.warn(TRUE, "just a true seen")
}
