% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resetpar.R
\name{resetpar}
\alias{resetpar}
\title{resetpar}
\usage{
resetpar(oldpar)
}
\arguments{
\item{oldpar}{graphical parameters}
}
\value{
nothing
}
\description{
Resets the par if necessary.
}
\examples{
# no examples
}
