% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopif.R
\name{stopif}
\alias{stopif}
\title{stopif}
\usage{
stopif(cond, txt)
}
\arguments{
\item{cond}{logical: condition to test}

\item{txt}{character: error message}
}
\value{
nothing
}
\description{
A equivalent to \code{stopifnot}: if \code{cond} is \code{TRUE} then a error is thrown.
}
\examples{
if (interactive()) stopif(1+1==2, "1+1 can not be 2, this is fake science!")
}
