% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table2dataframe.R
\name{table2dataframe}
\alias{table2dataframe}
\title{table2dataframe}
\usage{
table2dataframe(tab, ...)
}
\arguments{
\item{tab}{table: contingency table}

\item{...}{further parameters given to \link[base:table]{base::as.data.frame.table}}
}
\value{
a data frame with \code{sum(tab)} rows and \code{length(dim(tab))} cols
}
\description{
Converts a table to a full data frame.
}
\examples{
table2dataframe(Titanic)
}
