% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toRDS.R
\name{toRDS}
\alias{toRDS}
\title{toRDS}
\usage{
toRDS(...)
}
\arguments{
\item{...}{data sets to save}
}
\value{
returns the name of the created files
}
\description{
Saves one or more data sets in RDS format to a temporary directory (\code{tmpdir()}).
Data sets must have the class \code{ts} or something that can be converted
to a data frame, e.g. \code{matrix}, \code{table}, etc.
}
\examples{
toRDS(Titanic) # saves to tempdir/Titanic.rds
}
