% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucfirst.R
\name{ucfirst}
\alias{ucfirst}
\title{ucfirst}
\usage{
ucfirst(txt)
}
\arguments{
\item{txt}{character:}
}
\value{
character
}
\description{
Uppercases the first character in \code{txt}.
}
\examples{
ucfirst("hello world")
}
